Move TestingResult to separate file
